+++ src/testing.rs
@@ -22,7 +22,7 @@ pub fn subtest_search_point(board: &Array2::<u8>, gen_pattern: &Array1::<u8>, co
     let expected_matches: Vec<Match> = Vec::from([(start, end)]);
     for x in 0..SIDE_LEN {
         for y in 0..SIDE_LEN {
-            let point = (x as i8, y as i8);
+            let point = (x as isize, y as isize);
             let matches = search_point(board, gen_pattern, color, point);
 
             if point_is_on_line(point, start, end, true) {
@@ -35,11 +35,11 @@ pub fn subtest_search_point(board: &Array2::<u8>, gen_pattern: &Array1::<u8>, co
 }
 
 pub fn subtest_search_point_own(board: &Array2::<u8>, gen_pattern: &Array1::<u8>, color: u8,
-                                own_sqs: &Array1<i8>, start: Point, end: Point) {
+                                own_sqs: &Array1<isize>, start: Point, end: Point) {
     let expected_matches: Vec<Match> = Vec::from([(start, end)]);
     for x in 0..SIDE_LEN {
         for y in 0..SIDE_LEN {
-            let point = (x as i8, y as i8);
+            let point = (x as isize, y as isize);
             let matches = search_point_own(board, gen_pattern, color, point, own_sqs);
             if board[(x, y)] == color && point_is_on_line(point, start, end, true) {
                 assert!(matches_are_equal(&matches, &expected_matches));
@@ -51,7 +51,7 @@ pub fn subtest_search_point_own(board: &Array2::<u8>, gen_pattern: &Array1::<u8>
 }
 
 pub fn subtest_search_board_next_sq(board: &mut Array2::<u8>, gen_pattern: &Array1::<u8>, color: u8,
-                                    own_sqs: &Array1<i8>, defcon: i8, start: Point, end: Point) {
+                                    own_sqs: &Array1<isize>, defcon: usize, start: Point, end: Point) {
     for own_sq in own_sqs {
         let test_sq = point_on_line(start, end, *own_sq);
         let expected_ns_matches: Vec<NSQMatch> = Vec::from([(test_sq, (start, end))]);
@@ -63,7 +63,7 @@ pub fn subtest_search_board_next_sq(board: &mut Array2::<u8>, gen_pattern: &Arra
 
         assert!(next_sq_matches_are_subset(&expected_ns_matches, &ns_matches));
 
-        if WIN_LENGTH - defcon as usize > 2 {
+        if WIN_LENGTH - defcon > 2 {
             for nsm in ns_matches {
                 assert!(point_is_on_line(nsm.0, start, end, false));
             }
@@ -72,14 +72,14 @@ pub fn subtest_search_board_next_sq(board: &mut Array2::<u8>, gen_pattern: &Arra
 }
 
 pub fn subtest_search_point_next_sq(board: &mut Array2::<u8>, gen_pattern: &Array1::<u8>, color: u8,
-                                    own_sqs: &Array1<i8>, defcon: i8, start: Point, end: Point) {
+                                    own_sqs: &Array1<isize>, defcon: usize, start: Point, end: Point) {
     for own_sq in own_sqs {
         let test_sq = point_on_line(start, end, *own_sq);
         let expected_ns_matches: Vec<NSQMatch> = Vec::from([(test_sq, (start, end))]);
 
         for x in 0..SIDE_LEN {
             for y in 0..SIDE_LEN {
-                let point = (x as i8, y as i8);
+                let point = (x as isize, y as isize);
 
                 let stored_val = board[(test_sq.0 as usize, test_sq.1 as usize)];
                 board[(test_sq.0 as usize, test_sq.1 as usize)] = EMPTY;
@@ -90,12 +90,12 @@ pub fn subtest_search_point_next_sq(board: &mut Array2::<u8>, gen_pattern: &Arra
                     assert!(next_sq_matches_are_subset(&expected_ns_matches, &ns_matches));
                 } else if point_is_on_line(point, start, end, false) {
                 } else {
-                    if WIN_LENGTH - defcon as usize > 2 {
+                    if WIN_LENGTH - defcon > 2 {
                         assert_eq!(ns_matches.len(), 0);
                     }
                 }
 
-                if WIN_LENGTH - defcon as usize > 2 {
+                if WIN_LENGTH - defcon > 2 {
                     for nsm in ns_matches {
                         assert!(point_is_on_line(nsm.0, start, end, false));
                     }
@@ -106,14 +106,14 @@ pub fn subtest_search_point_next_sq(board: &mut Array2::<u8>, gen_pattern: &Arra
 }
 
 pub fn subtest_search_point_own_next_sq(board: &mut Array2::<u8>, gen_pattern: &Array1::<u8>, color: u8,
-                                    own_sqs: &Array1<i8>, defcon: i8, start: Point, end: Point) {
+                                    own_sqs: &Array1<isize>, defcon: usize, start: Point, end: Point) {
     for own_sq in own_sqs {
         let test_sq = point_on_line(start, end, *own_sq);
         let expected_ns_matches: Vec<NSQMatch> = Vec::from([(test_sq, (start, end))]);
 
         for x in 0..SIDE_LEN {
             for y in 0..SIDE_LEN {
-                let point = (x as i8, y as i8);
+                let point = (x as isize, y as isize);
 
                 let stored_val = board[(test_sq.0 as usize, test_sq.1 as usize)];
                 board[(test_sq.0 as usize, test_sq.1 as usize)] = EMPTY;
@@ -125,12 +125,12 @@ pub fn subtest_search_point_own_next_sq(board: &mut Array2::<u8>, gen_pattern: &
                     assert!(next_sq_matches_are_subset(&expected_ns_matches, &ns_matches));
                 } else if point_is_own_sq && point_is_on_line(point, start, end, false) {
                 } else {
-                    if WIN_LENGTH - defcon as usize > 2 {
+                    if WIN_LENGTH - defcon > 2 {
                         assert_eq!(ns_matches.len(), 0);
                     }
                 }
 
-                if WIN_LENGTH - defcon as usize > 2 {
+                if WIN_LENGTH - defcon > 2 {
                     for nsm in ns_matches {
                         assert!(point_is_on_line(nsm.0, start, end, false));
                     }
@@ -140,7 +140,7 @@ pub fn subtest_search_point_own_next_sq(board: &mut Array2::<u8>, gen_pattern: &
     }
 }
 
-pub fn subtest_search_fns(gen_pattern: &Array1::<u8>, color: u8, own_sqs: &Array1<i8>, defcon: i8) {
+pub fn subtest_search_fns(gen_pattern: &Array1::<u8>, color: u8, own_sqs: &Array1<isize>, defcon: usize) {
     let pattern = get_pattern(gen_pattern, color);
     let length = pattern.len();
 
@@ -148,10 +148,10 @@ pub fn subtest_search_fns(gen_pattern: &Array1::<u8>, color: u8, own_sqs: &Array
         for j in 0..SIDE_LEN {
             for d in 0..NUM_DIRECTIONS {
                 let mut board = new_board();
-                if apply_pattern(&mut board, &pattern, (i as i8, j as i8), d) {
-                    let (row_inc, col_inc) = increments(d as i8);
-                    let start = (i as i8, j as i8);
-                    let end = (idx(i as i8, row_inc, length - 1), idx(j as i8, col_inc, length - 1));
+                if apply_pattern(&mut board, &pattern, (i as isize, j as isize), d) {
+                    let (row_inc, col_inc) = increments(d as isize);
+                    let start = (i as isize, j as isize);
+                    let end = (idx(i as isize, row_inc, length - 1), idx(j as isize, col_inc, length - 1));
 
                     subtest_search_board(&board, gen_pattern, color, start, end);
                     subtest_search_point(&board, gen_pattern, color, start, end);
